xds refactor
